event-registration-master/requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.4
django-cors-headers==3.3.0
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
numpy==1.18.4
phonenumbers==8.12.4
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1